+++ src/components/PolygonMapTesting.jsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useRef } from "react";
 import H from '@here/maps-api-for-javascript'; // npm install @here/maps-api-for-javascript --registry=https://repo.platform.here.com/artifactory/api/npm/maps-api-for-javascript/
 import polygonmapAPI from "../utils/polygonmapAPI";
-import PolygonObject from "./PolygonObject";
+import PolygonObject from "./PolygonObjectTesting";
 
-const Map = () => {
+const PolygonMapTesting = () => {
   // console.log("APIkey of here map: ", polygonmapAPI);
   // console.log(props);
   const mapRef = useRef(null);
@@ -76,7 +76,7 @@ return (map)
       // </>);
 }
 
-export default Map;
+export default PolygonMapTesting;
 
 // https://www.here.com/docs/bundle/maps-api-for-javascript-developer-guide/page/topics/react-practices.html#add-a-react-functional-component-that-displays-a-map
 // https://developer.here.com/documentation/examples/maps-js/geoshapes/polygon-on-the-map
